+++ core/src/component/context.rs
@@ -287,13 +287,20 @@ where
     }
 
     /// Returns a handle to the Kompact system this component is a part of
-    pub fn system(&self) -> ContextSystemHandle {
+    #[cfg(not(feature = "distributed"))]
+    pub fn system(&self) -> impl SystemHandle {
+        self.context_system()
+    }
+
+    /// Returns a handle to the Kompact system this component is a part of
+    #[cfg(feature = "distributed")]
+    pub fn system(&self) -> impl DistributedSystemHandle + SystemHandle {
         self.context_system()
     }
 
     /// Returns a reference to the system dispatcher
     pub fn dispatcher_ref(&self) -> DispatcherRef {
-        self.system().dispatcher_ref()
+        self.context_system().dispatcher_ref()
     }
 
     /// Returns a reference to the system's deadletter box
