+++ network/Cargo.toml
@@ -16,20 +16,19 @@ repository = "https://github.com/kompics/kompact"
 license = "MIT"
 
 [features]
-default = ["distributed"]
-distributed = ["kompact/distributed"]
+default = []
 low_latency = ["kompact/low_latency"]
 
 [dependencies]
-kompact = { version = "0.11.3", path = "../core" }
-arc-swap = "1.9"
-async-std = "1.13"
-bytes = "1.11"
-crossbeam-channel = "0.5"
-futures = "0.3"
-ipnet = "2.12"
+kompact = { version = "0.11.3", path = "../core", features = ["distributed"] }
+arc-swap = { workspace = true }
+async-std = { workspace = true }
+bytes = { workspace = true }
+crossbeam-channel = { workspace = true }
+futures = { workspace = true }
+ipnet = { workspace = true }
 iprange = "0.6"
-lru = "0.16"
+lru = { workspace = true }
 mio = { version = "1", features = ["net", "os-poll"] }
-rustc-hash = "2"
-uuid = { version = "1.23", features = ["v4"] }
+rustc-hash = { workspace = true }
+uuid = { workspace = true }
